+++ task/Tasks.js
@@ -1,6 +1,5 @@
 "use strict";
 /// <reference path="./task-results.d.ts" />
-var Q = require("q");
 var Task = require("./Task");
 var TaskSet = require("./TaskSet");
 var Tasks;
@@ -8,13 +7,12 @@ var Tasks;
     /** Create a task
      * @param name the name of the task
      * @param action the unit of work performed by this task, a function or promise
-     * @param [dfd] an optional Deferred to use for tracking the completion/failure of the task, if not provided a default will be created using 'Q.defer()'
+     * @param dfd IDeferred for tracking the completion/failure of the task, if not provided a default will be created using 'Q.defer()'
      */
-    function newTask(name, action, dfd) {
-        if (dfd === void 0) { dfd = Q.defer(); }
-        return new Task(name, action, dfd);
+    function startTask(name, action) {
+        return Task.startTask(name, action);
     }
-    Tasks.newTask = newTask;
+    Tasks.startTask = startTask;
     /** Create a task set
      * @param name the name of the task set
      * @param action the unit of work performed by this task, a function or promise