+++ task/Tasks.ts
@@ -1,5 +1,4 @@
 ﻿/// <reference path="./task-results.d.ts" />
-import Q = require("q");
 import Task = require("./Task");
 import TaskSet = require("./TaskSet");
 
@@ -8,10 +7,10 @@ module Tasks {
     /** Create a task
      * @param name the name of the task
      * @param action the unit of work performed by this task, a function or promise 
-     * @param [dfd] an optional Deferred to use for tracking the completion/failure of the task, if not provided a default will be created using 'Q.defer()'
+     * @param dfd IDeferred for tracking the completion/failure of the task, if not provided a default will be created using 'Q.defer()'
      */
-    export function newTask<R1, S1>(name: string, action: (() => R1) | Q.IPromise<R1>, dfd: Q.Deferred<R1> = Q.defer<R1>()): TaskResults.Task<R1, S1> {
-        return new Task<R1, S1>(name, action, dfd);
+    export function startTask<R1, S1>(name: string, action: PsPromise<R1, S1>): TaskResults.Task<R1, S1> {
+        return Task.startTask(name, action);
     }