+++ test/TaskTest.js
@@ -1,14 +1,15 @@
 "use strict";
 Object.defineProperty(exports, "__esModule", { value: true });
 var chai = require("chai");
+var Q = require("q");
 var Tasks = require("../task/Tasks");
 var asr = chai.assert;
 suite("Task", function TaskTest() {
     test("task-success", function taskSuccessTest(done) {
         // test success
-        var res1 = Tasks.newTask("a", function () { return "a-success"; });
+        var res1 = Tasks.startTask("a", Q.resolve("a-success"));
         asr.equal(res1.name, "a");
-        asr.equal(res1.state, "CREATED");
+        asr.equal(res1.state, "RUNNING");
         res1.getPromise().then(function (res) {
             asr.equal(res, "a-success");
             asr.equal(res1.getResult(), "a-success");
@@ -18,13 +19,12 @@ suite("Task", function TaskTest() {
             asr.equal(true, false, "task " + res1.name + " failed unexpectedly");
             done();
         });
-        res1.start();
     });
     test("task-failure", function taskFailureTest(done) {
         // test failure
-        var res1 = Tasks.newTask("a", function () { throw "a-error"; });
+        var res1 = Tasks.startTask("a", Q.reject("a-error"));
         asr.equal(res1.name, "a");
-        asr.equal(res1.state, "CREATED");
+        asr.equal(res1.state, "RUNNING");
         res1.getPromise().then(function (res) {
             asr.equal(true, false, "task " + res1.name + " succeeded unexpectedly");
             done();
@@ -34,6 +34,5 @@ suite("Task", function TaskTest() {
             asr.equal(res1.state, "ERRORED");
             done();
         });
-        res1.start();
     });
 });