+++ test/TaskTest.ts
@@ -1,18 +1,18 @@
 ﻿import chai = require("chai");
 import mocha = require("mocha");
+import Q = require("q");
 import Tasks = require("../task/Tasks");
 
 var asr = chai.assert;
-
 
 suite("Task", function TaskTest() {
 
     test("task-success", function taskSuccessTest(done) {
         // test success
-        var res1 = Tasks.newTask("a", () => "a-success");
+        var res1 = Tasks.startTask("a", <any>Q.resolve("a-success"));
 
         asr.equal(res1.name, "a");
-        asr.equal(res1.state, "CREATED");
+        asr.equal(res1.state, "RUNNING");
 
         res1.getPromise().then((res) => {
             asr.equal(res, "a-success");
@@ -23,17 +23,15 @@ suite("Task", function TaskTest() {
             asr.equal(true, false, "task " + res1.name + " failed unexpectedly");
             done();
         });
-
-        res1.start();
     });
 
 
     test("task-failure", function taskFailureTest(done) {
         // test failure
-        var res1 = Tasks.newTask("a", () => { throw "a-error"; });
+        var res1 = Tasks.startTask("a", <any>Q.reject("a-error"));
 
         asr.equal(res1.name, "a");
-        asr.equal(res1.state, "CREATED");
+        asr.equal(res1.state, "RUNNING");
 
         res1.getPromise().then((res) => {
             asr.equal(true, false, "task " + res1.name + " succeeded unexpectedly");
@@ -44,8 +42,6 @@ suite("Task", function TaskTest() {
             asr.equal(res1.state, "ERRORED");
             done();
         });
-
-        res1.start();
     });
 
 });